Use append mode when writing data with `--ner_output_file`
Because the pre-commit hook can invoke the program multiple times, we were
overwriting NER output data when the final invocation wrote to the file.

Change to using append mode, which means NER output data from mulitple
invocations will all be kept. However this also means the program will
never truncate the file, so the user may want to do that manually.

The `NoExcludeFilePassedException` exception stuff was removed because the
regular behaviour of failing when trying to write to `None` is fine, and
there's already a guard against trying to write the NER output file with
no destination specified.